Added sitemap infrastructure for creating iCal calendars from meeting memos. The sitemap will match links like:
/doc/admin/weekly/2006/Tasks_2006-06-19_Sjur.ics

and return an iCal file.

WARNING! The returned file is INVALID, and can't be processed by iCal and other calendar applications. The xsl used to extract and produce the iCal file is just a template, and needs substantial work. But the framework is there, and linking works, now we *only* have to fix the xsl processing:-)