Added entity for pdf-config-file (used by our pdf setup to get Sámi chars correct in pdf).
The entity contains the absolute path to the config file, and is user dependent.

That is, EVERYBODY needs to CHANGE this entity to suit their local environment (at least if you need local pdf support by Forrest).

The entity is not yet used, but is needed for recent changes to Forrest I have done so that we can use plain-vanilla Forrest instead of the tailored one Børre has made.