Initial support for automatic deploying of our own plugins - it doesn't yet work, though.
Also added the proofing tools test result input plugin to the list of required plugins. The plugin description is found, and forrest thus runs ok, but the plugin is not automatically deployed, which renders the relevant pages unreachable by forrest. I'm working on it (bug #939).

Also added (as comments) a couple of other useful plugins, so that they can easily be commented in when needed.

Finally I added support for a user-independent specification of the command-line prefs (again as a comment - we don't use it in normal operation). Not tested - it might not work...