These are all the javascript files generated by the Divvun graphing module. They need to be stored here, and added to svn, to be available to Forrest and for building static Forrest pages.
One additional file is built by the graphing module: index.html. That file should NOT be part of svn, and svn does now ignore it. The file is useful for debugging the graph rendering. It should automatically display all the graphs generated. Just open it in your favourite web browser.