Added processing of sub-sections and sub-subsections, corrected processing of first and non-first top-level sections. Made ordered lists into unordered lists with incremental display.
With this commit we have a quite decent tool to produce HTML-based slide-shows directly from jspwiki files. If jspwiki is too limited, the forrest xml document format can also be used (but support for it is presently not as good as it should be). For more details on W3C Slidy, see http://www.w3.org/Talks/Tools/Slidy2.