First version of a local modification to the Chaperon grammar and stylesheet used for jspwiki formatting. The enhancements are: * support for formatting, links and breaks within table cells * initial support for definition lists - only one term + definition is allowed in this version
NB. Because of a strange bug(?) in the Chaperon parser, the correct definition list delimiters as defined by the jspwiki grammar is not accepted. Instead the marks #; and ##: are used, and they must be on consecutive lines, not the same line.

NB2. Because of a small bug in the wiki plugin in Forrest, the local modifications will not be picked up by Forrest automatically. The plan is to fix the bug, and release a new version of the wiki plugin. Those interested can contact me for details on the required changes.

Chaperon is used by Cocoon, which is used by Forrest. Chaperon has its homepage at chaperon.sf.net.