Commented out test elements.
Found a BUG in Forrest - it seems that someone forgot to add a test that is used for level1 in the level2 check. When added, multiple level2 tab highlighting is gone, eand everything seems to work as it should. NB! Not tested enough, there might be side effects we haven't seen yet!